rag-chatbot/requirements.txt
langchain
python-dotenv
unstructured
markdown
langchain-google-genai>=1.0.0
google-generativeai>=0.8.0
langchain-community
langchain-core
faiss-cpu
fastapi
uvicorn
pypdf
pydantic
langchain-classic
python-multipart